refactored: normalized directory structure
