+++ src/renderer/res/trash.jpg
