+++ src/renderer/scripts/currentIndex.js
