+++ src/renderer/scripts/fileObjects.js
