+++ src/renderer/scripts/finalize.js
