+++ src/renderer/scripts/inspect.js
