+++ src/renderer/scripts/progress.js
