+++ src/renderer/scripts/rename.js
