+++ src/renderer/scripts/settings.js
