+++ src/renderer/scripts/swipe.js
