+++ src/renderer/scripts/tooltip.js
