+++ src/renderer/scripts/trashModal.js
